Change some uses of long to more appropriate types
* glom/libglom/python_embed/py_glom_record.[cc|h]:
* glom/libglom/python_embed/py_glom_related.[cc|h]:
* glom/libglom/python_embed/py_glom_relatedrecord.[cc|h]:
* glom/libglom/utils.cc:
* glom/utility_widgets/db_addel/glom_db_treemodel.cc:
* glom/utils_ui.cc: Use other types than long where appropriate.

+++ glom/utils_ui.cc
@@ -280,7 +280,7 @@ Glib::RefPtr<Gdk::Pixbuf> Utils::get_pixbuf_for_gda_value(const Gnome::Gda::Valu
         guint8* puiData = (guint8*)buffer_binary;
         try
         {
-          refPixbufLoader->write(puiData, (glong)buffer_binary_length);
+          refPixbufLoader->write(puiData, static_cast<gsize>(buffer_binary_length));
           result = refPixbufLoader->get_pixbuf();
 
           refPixbufLoader->close(); //This throws if write() threw, so it must be inside the try block.
